+++ src/app/lib/device/dlna.js
@@ -1,10 +1,9 @@
 (function (App) {
     'use strict';
-    var Browser = require('nodecast-js');
-    var MediaRendererClient = require('upnp-mediarenderer-client');
+    var dlnacasts = require('dlnacasts')();
     var xmlb = require('xmlbuilder');
     var collection = App.Device.Collection;
-    var browser = new Browser();
+
 
     var makeID = function (baseID) {
         return 'dlna-' + baseID.replace('-', '');
@@ -18,136 +17,124 @@
         makeID: makeID,
 
         initialize: function (attrs) {
-            this.device = attrs.device;
-            this.client = new MediaRendererClient(this.device.xml);
-            this.attributes.name = this.device.name;
-            this.attributes.address = this.device.host;
+            this.player = attrs.player;
+            this.attributes.name = this.player.name;
+            this.attributes.address = this.player.host;
         },
 
         play: function (streamModel) {
             var url = streamModel.get('src');
+            var self = this;
+            var media;
             var url_video = url;
             var url_subtitle = 'http:' + url.split(':')[1] + ':9999/video.srt';
             var metadata = null;
             var subtitle = streamModel.get('subFile');
             if (subtitle) {
-                metadata = xmlb.create('DIDL-Lite', {
-                        'headless': true
-                    })
-                    .att({
-                        'xmlns': 'urn:schemas-upnp-org:metadata-1-0/DIDL-Lite/',
-                        'xmlns:dc': 'http://purl.org/dc/elements/1.1/',
-                        'xmlns:upnp': 'urn:schemas-upnp-org:metadata-1-0/upnp/',
-                        'xmlns:dlna': 'urn:schemas-dlna-org:metadata-1-0/',
-                        'xmlns:sec': 'http://www.sec.co.kr/',
-                        'xmlns:xbmc': 'urn:schemas-xbmc-org:metadata-1-0/'
-                    })
-                    .ele('item', {
-                        'id': '0',
-                        'parentID': '-1',
-                        'restricted': '1'
-                    })
-                    .ele('dc:title', {}, streamModel.get('title'))
-                    .insertAfter('res', {
-                        'protocolInfo': 'http-get:*:video/mp4:*',
-                        'xmlns:pv': 'http://www.pv.com/pvns/',
-                        'pv:subtitleFileUri': url_subtitle,
-                        'pv:subtitleFileType': 'srt'
-                    }, url_video)
-                    .insertAfter('res', {
-                        'protocolInfo': 'http-get:*:text/srt:'
-                    }, url_subtitle)
-                    .insertAfter('res', {
-                        'protocolInfo': 'http-get:*:smi/caption'
-                    }, url_subtitle)
-                    .insertAfter('sec:CaptionInfoEx', {
-                        'sec:type': 'srt'
-                    }, url_subtitle)
-                    .insertAfter('sec:CaptionInfo', {
-                        'sec:type': 'srt'
-                    }, url_subtitle)
-                    .insertAfter('upnp:class', {}, 'object.item.videoItem.movie')
-                    .end({
-                        pretty: false
-                    });
+                media = {
+                    title: streamModel.get('title'),
+                    subtitles: ['http:' + url.split(':')[1] + ':9999/subtitle.vtt'],
+                };
             } else {
-                metadata = xmlb.create('DIDL-Lite', {
-                        'headless': true
-                    })
-                    .att({
-                        'xmlns': 'urn:schemas-upnp-org:metadata-1-0/DIDL-Lite/',
-                        'xmlns:dc': 'http://purl.org/dc/elements/1.1/',
-                        'xmlns:upnp': 'urn:schemas-upnp-org:metadata-1-0/upnp/',
-                        'xmlns:dlna': 'urn:schemas-dlna-org:metadata-1-0/',
-                        'xmlns:sec': 'http://www.sec.co.kr/',
-                        'xmlns:xbmc': 'urn:schemas-xbmc-org:metadata-1-0/'
-                    })
-                    .ele('item', {
-                        'id': '0',
-                        'parentID': '-1',
-                        'restricted': '1'
-                    })
-                    .ele('dc:title', {}, streamModel.get('title'))
-                    .insertAfter('res', {
-                        'protocolInfo': 'http-get:*:video/mp4:*',
-                    }, url_video)
-                    .insertAfter('upnp:class', {}, 'object.item.videoItem.movie')
-                    .end({
-                        pretty: false
-                    });
+                media = {
+                    title: streamModel.get('title')
+                };
             }
-            this.client.load(url_video, {
-                metadata: metadata,
-                autoplay: true,
+            win.info('DLNA: play ' + url + ' on \'' + this.get('name') + '\'');
+            win.info('DLNA: connecting to ' + this.player.host);
 
-            }, function (err, result) {
-                if (err) {
-                    throw err;
-                }
-
+            self.player.play(url_video, media , function (err, status) {
+              if (err) {
+                win.error('DLNA.play error: ', err);
+              } else {
+                win.info('Playing ' + url + ' on ' + self.get('name'));
+                self.set('loadedMedia', status.media);
+              }
             });
-        },
+    this.player.on('status', function (status) {
+                  self._internalStatusUpdated(status);
+              });
+          },
 
         stop: function () {
-            this.client.stop();
+            this.player.stop();
         },
 
         pause: function () {
-            this.client.pause();
+            this.player.pause();
         },
 
         forward: function () {
-            this.client.seek(30);
+            this.player.seek(30);
         },
 
         backward: function () {
-            this.client.seek(-30);
+            this.player.seek(-30);
+        },
+
+        seek: function (seconds) {
+            win.info('DLNA: seek %s', seconds);
+            this.get('player').seek(seconds, function (err, status) {
+                if (err) {
+                    win.error('DLNA.seek:Error', err);
+                }
+            });
+        },
+        seekTo: function (newCurrentTime) {
+            win.info('DLNA: seek to %ss', newCurrentTime);
+            this.get('player').seek(newCurrentTime, function (err, status) {
+                if (err) {
+                    win.error('DLNA.seek:Error', err);
+                }
+            });
+        },
+
+        seekPercentage: function (percentage) {
+            win.info('DLNA: seek percentage %s%', percentage.toFixed(2));
+            var newCurrentTime = this.player._status.duration / 100 * percentage;
+            this.seekTo(newCurrentTime.toFixed());
         },
 
         unpause: function () {
-            this.client.play();
+            this.player.play();
+        },
+        updateStatus: function () {
+            var self = this;
+            this.get('player').status(function (err, status) {
+                if (err) {
+                    return win.info('DLNA.updateStatus:Error', err);
+                }
+                self._internalStatusUpdated(status);
+            });
+        },
+
+        _internalStatusUpdated: function (status) {
+          if (status  === undefined) {
+              status = this.player._status;
+          }
+            // If this is the active device, propagate the status event.
+            if (collection.selected.id === this.id) {
+                App.vent.trigger('device:status', status);
+            }
         }
     });
 
 
-    browser.onDevice(function (device) {
-        device.onError(function (err) {
-            win.error('DNLA device error', err);
-        });
-
+    dlnacasts.on('update', function (player) {
         if (collection.where({
-                id: device.host
+                id: player.host
             }).length === 0) {
-            win.info('Found DLNA Device: %s at %s', device.name, device.host);
+            win.info('Found DLNA Device: %s at %s', player.name, player.host);
             collection.add(new Dlna({
-                id: device.host,
-                device: device
+                id: player.host,
+                player: player
             }));
         }
     });
 
     win.info('Scanning: Local Network for DLNA devices');
-    browser.start();
+    dlnacasts.update();
+
 
     App.Device.Dlna = Dlna;
 })(window.App);